+++ stories/components/Modal/ModalWithForm.vue
@@ -2,7 +2,7 @@
 	<div>
 		<a v-on:click="modalIsActive = !modalIsActive">Open Modal with a form</a>
 
-		<modal :active="modalIsActive" :title="title">
+		<modal :active="modalIsActive" :title="title" defaultStyles>
 			<form slot="content" class="example-form">
 				<p class="example-form-feedback" v-show="feedback.length">{{ feedback }}</p>
 
@@ -26,7 +26,7 @@
 				</div>
 			</form>
 			<div slot="footer">
-				<btn kind="secondary" @click="closeModal">Close</btn>
+				<btn @click="closeModal">Close</btn>
 				<btn kind="primary" @click="submitForm">Send</btn>
 			</div>
 		</modal>