+++ stories/components/Modal/ModalWithInfo.vue
@@ -2,7 +2,15 @@
 	<div>
 		<a v-on:click="modalIsActive = !modalIsActive">Open Info Modal</a>
 
-		<modal :active="modalIsActive" :title="title" :close-icon="true" :close-on-outside-click="true" :close-on-escape-press="true" @closed="closeModal">
+		<modal
+			:active="modalIsActive"
+			:title="title"
+			defaultStyles
+			close-icon
+			close-on-outside-click
+			close-on-escape-press
+			@closed="closeModal"
+		>
 			<p slot="content">Lorem ipsum dolor sit amet consectetur adipisicing elit. Numquam cupiditate natus nostrum? Esse error dolor suscipit. Animi, eveniet, inventore magnam ratione molestiae vitae quam omnis rem nulla dicta veniam similique.<br>
 			Lorem ipsum dolor sit amet consectetur adipisicing elit. Numquam cupiditate natus nostrum? Esse error dolor suscipit. Animi, eveniet, inventore magnam ratione molestiae vitae quam omnis rem nulla dicta veniam similique.</p>
 		</modal>