+++ stories/components/Modal/ModalWithLongText.vue
@@ -2,7 +2,7 @@
 	<div>
 		<a v-on:click="modalIsActive = !modalIsActive">Open Modal with lots of text</a>
 
-		<modal :active="modalIsActive" :title="title">
+		<modal :active="modalIsActive" :title="title" defaultStyles>
 			<p slot="content">
 				Lorem ipsum dolor sit amet consectetur adipisicing elit. Numquam cupiditate natus nostrum? Esse error dolor suscipit. Animi, eveniet, inventore magnam ratione molestiae vitae quam omnis rem nulla dicta veniam similique.
 				Lorem ipsum dolor sit amet consectetur adipisicing elit. Numquam cupiditate natus nostrum? Esse error dolor suscipit. Animi, eveniet, inventore magnam ratione molestiae vitae quam omnis rem nulla dicta veniam similique.
@@ -13,7 +13,7 @@
 				Lorem ipsum dolor sit amet consectetur adipisicing elit. Numquam cupiditate natus nostrum? Esse error dolor suscipit. Animi, eveniet, inventore magnam ratione molestiae vitae quam omnis rem nulla dicta veniam similique.
 			</p>
 			<div slot="footer">
-				<btn kind="secondary" @click="closeModal">Close</btn>
+				<btn @click="closeModal">Close</btn>
 				<btn kind="primary">Send</btn>
 			</div>
 		</modal>